refactor(handler): adopt 32bit int type in request (#417)
Because

- the grpc-gateway will convert `int64` to `string`

This commit

- use `int32` for `page_size`, `total_size` and `file_length` to avoid
this problem

+++ pkg/handler/private_handler.go
@@ -47,7 +47,7 @@ func (h *PrivateHandler) ListModelsAdmin(ctx context.Context, req *modelPB.ListM
 	resp := modelPB.ListModelsAdminResponse{
 		Models:        pbModels,
 		NextPageToken: nextPageToken,
-		TotalSize:     totalSize,
+		TotalSize:     int32(totalSize),
 	}
 
 	return &resp, nil
